* Blits are now transparent and so is hit detection

+++ ui/freeamp/beos/src/CanvasView.cpp
@@ -77,8 +77,8 @@ CanvasView::MouseDown( BPoint p )
     ConvertToScreen( &p );
 
     Pos pos;
-    pos.x = p.x;
-    pos.y = p.y;
+    pos.x = (int)p.x;
+    pos.y = (int)p.y;
     m_parent->HandleMouseLButtonDown( pos );
 }
